refactor(protest_verification): detach from protest package
